refactor move protocol_server to its own package
