+++ p2p/service/service.go
@@ -17,3 +17,31 @@ type Service interface {
 	SendMessage(nodeID string, protocol string, payload []byte) error
 	Shutdown()
 }
+
+type Data interface {
+	messageData()
+	Bytes() []byte
+}
+
+type Data_Bytes struct {
+	Payload []byte
+}
+
+type Data_MsgWrapper struct {
+	Req     bool
+	MsgType uint32
+	ReqID   uint64
+	Payload []byte
+}
+
+func (m Data_Bytes) messageData() {}
+
+func (m Data_Bytes) Bytes() []byte {
+	return m.Payload
+}
+
+func (m Data_MsgWrapper) messageData() {}
+
+func (m Data_MsgWrapper) Bytes() []byte {
+	return m.Payload
+}
